+++ leap_ec/decoder.py
@@ -64,13 +64,13 @@ class Decoder(abc.ABC):
     evaluate its own fitness:
 
     >>> ind.evaluate()
-    836.4453949...
+    np.float64(836.4453949...)
 
     Calling `evaluate()` also has the side effect of setting the fitness
     attribute:
 
     >>> ind.fitness
-    836.4453949...
+    np.float64(836.4453949...)
 
     """
